Fix bug when documentclass was amsart. This was because sections that appeared as \toc\(sub\)*section in the aux file were not being parsed correctly by ReadTOC in motion.vim
I also changed it so that it would only add entries to the TOC that we
knew how to parse. This changes its previous behavior, where things
sections that appeared as \tocpart in the aux file would be added to the TOC.

Removed some substitutions for secnum and text. The title text is not as
pleasing, but it's complete.

Fix search bug in latextoc.vim when `$ $` equations were present in the
title str. I changed this in s:EscapeTitle in latextoc.vim. For example

    \section{Test $asd_{12}$ $\mathcal{b}$}

This method of escaping will certainly produce duplicate matches in some
scenarios.